+++ lib/activerecord_cached.rb
@@ -8,6 +8,7 @@ require_relative "activerecord_cached/cache"
 require_relative "activerecord_cached/configuration"
 require_relative "activerecord_cached/limit_checks"
 require_relative "activerecord_cached/railtie" if defined?(::Rails::Railtie)
+require_relative "activerecord_cached/redis_mutex"
 require_relative "activerecord_cached/version"
 
 module ActiveRecordCached