+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## [Unreleased]
 
+## Added
+
+- cache expiry and race_condition_ttl
+
+## Fixed
+
+- some race conditions that would affect cache clearing by CRUD operations
+
 ## [1.0.0] - 2024-01-23
 
 - Initial release